backend/requirements.txt
cachetools
fastapi[standard]
httpx
pydantic
pydantic-settings
python-jose[cryptography]
uvicorn
